requirements.txt
mkdocs==1.6.1
mkdocs-material==9.5.34
requests==2.32.3
PyYAML==6.0.2
Jinja2==3.1.4
python-frontmatter==1.1.0
feedparser==6.0.11
mkdocs-rss-plugin==1.12.1
